requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
python-dotenv==1.0.0
Werkzeug==3.0.1
python-dateutil>=2.8